refactor: replace FloorPlanWidget with SimpleMapWidget and update imports and added new stuff
